events_derived: merge into existing files
Move the derivation pipeline stage to events_pipeline
Move the derivation table initialization to tracee.go
